requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.2.26
langchain-openai==1.1.12
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.26
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0